cli: extract actions
